% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.ia.iterate.R
\name{F.ia.iterate}
\alias{F.ia.iterate}
\title{F.ia.iterate}
\usage{
F.ia.iterate(q, p, loc.y)
}
\arguments{
\item{q}{A vector of length K giving the last estimate for individual admixture of the test data.}

\item{p}{An K x 4 x M array containing the allele frequencies of the training data.}

\item{loc.y}{A 4 x K matrix containing the allele counts of the test data.}
}
\value{
A vector of length K, the new estimate for individual admixture of the test data.
}
\description{
A function iterated in order to estimate individual admixture.
}
