% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forenseq.EGDP.R
\docType{data}
\name{forenseq.EGDP}
\alias{forenseq.EGDP}
\title{Data for 56 AIMs from the Estonian Genome Diversity panel; this data can serve as test data}
\format{
An object of class \code{list} of length 2.
}
\source{
\href{http://mathgene.usc.es/snipper/Verogen\%20ForenSeq\%20DNA\%20Signature\%20Kit\%2056-AIM\%20Training\%20Set.xlsx}{Snipper homepage, EGDP sheet}
}
\usage{
data(forenseq.EGDP)
}
\description{
Data for 56 AIMs from the Estonian Genome Diversity panel; this data can serve as test data
}
\examples{
data(forenseq.EGDP)
}
\references{
Thermo Fisher Precision ID Ancestry panel
}
\keyword{datasets}
