% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forenseq.snipper.reference.R
\docType{data}
\name{forenseq.snipper.reference}
\alias{forenseq.snipper.reference}
\title{Data for 56 AIMs from the snipper.reference; this data can serve as the reference dataset}
\format{
An object of class \code{list} of length 2.
}
\source{
\href{http://mathgene.usc.es/snipper/Verogen\%20ForenSeq\%20DNA\%20Signature\%20Kit\%2056-AIM\%20Training\%20Set.xlsx}{Snipper homepage, first sheet}
}
\usage{
data(forenseq.snipper.reference)
}
\description{
Data for 56 AIMs from the snipper.reference; this data can serve as the reference dataset
}
\examples{
data(forenseq.snipper.reference)
}
\references{
Thermo Fisher Precision ID Ancestry panel
}
\keyword{datasets}
