% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.class.R
\name{get.class}
\alias{get.class}
\title{get.class}
\usage{
get.class(test.sample, training.sample, training.samLoc, method = "naiveBayes")
}
\arguments{
\item{test.sample}{An N_te x M matrix with entries AA,..., TT.}

\item{training.sample}{An N_Tr x M matrix with entries AA,..., TT.
From training.sample, allele frequencies are computed, and from test.data the individual admixture
is computed by iteration of F.ia.iterate until convergence.}

\item{training.samLoc}{A vector of length N giving the locations of all N samples. unique(training.samLoc) give}

\item{method}{The method used for classification. Until now, only a naive Bayes approach is implemented.}
}
\value{
An N x K matrix, where entry (i,k) is the probability that test-sample belongs to population k.
}
\description{
A function to obtain probabilities that a test sample belongs to a population/class. At the momenta, only a naive Bayesian classifier is implemented.
}
\examples{
training.sample = forenseq.snipper.reference$sample
training.samLoc = forenseq.snipper.reference$samLoc
test.sample = forenseq.EGDP$sample
get.class(test.sample, training.sample, training.samLoc)
}
