% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.freqs.R
\name{get.freqs}
\alias{get.freqs}
\title{Obtaining allele frequencies from allele count data}
\usage{
get.freqs(sample, samLoc)
}
\arguments{
\item{sample}{An N x M matrix with entries AA,..., TT. If N = 1, sample can also be a vector.}

\item{samLoc}{A vector of length N giving the locations of all N samples. unique(training.samLoc) gives all different geographical locations of the data. We set K = length(unique(training.samLoc)).}
}
\value{
A K x 4 x M array, where entry (k,i,m) is the allele frequency of allele i in population k at marker m.
}
\description{
Obtaining allele frequencies from allele count data
}
\examples{
sample = forenseq.snipper.reference$sample
samLoc = forenseq.snipper.reference$samLoc
get.freqs(sample, samLoc)

sample = VISAGE.BASIC.TOOL.snipper.reference$sample
samLoc = VISAGE.BASIC.TOOL.snipper.reference$samLoc
get.freqs(sample, samLoc)

}
