% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ia.R
\name{get.ia}
\alias{get.ia}
\title{get.ia}
\usage{
get.ia(test.sample, training.sample, training.samLoc, tol=1e-6, verbose = TRUE)
}
\arguments{
\item{test.sample}{An N x M matrix with entries AA,..., TT.}

\item{training.sample}{An N x M matrix with entries AA,..., TT.
If N = 1, test.sample can also be a vector.
From training.sample, allele frequencies are computed, and from test.data the individual admixture
is computed by iteration of F.ia.iterate until convergence.}

\item{training.samLoc}{A vector of length N giving the locations of all N samples. unique(training.samLoc) gives all different geographical locations of the data. We set K = length(unique(training.samLoc)).}

\item{tol}{The iteration stops if two consecutive calls of F.ia.iterate lead to vectors with L1-distance below tol.}

\item{verbose}{If TRUE, the number of iterations is reported, FALSE is silent mode}
}
\value{
An N x K matrix, where entry (i,k) is the estimated contribution of ancestry from population k for sample i.
}
\description{
Obtaining individual admixture from a sample, using a reference/training database
}
\examples{
training.sample = VISAGE.BASIC.TOOL.snipper.reference$sample
training.samLoc = VISAGE.BASIC.TOOL.snipper.reference$samLoc
test.sample = VISAGE.BASIC.TOOL.EGDP$sample[1:10,]
get.ia(test.sample, training.sample, training.samLoc, verbose = TRUE)
}
