% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.y.R
\name{get.y}
\alias{get.y}
\title{get.y}
\usage{
get.y(sample)
}
\arguments{
\item{sample}{A N x M matrix with entries AA,..., TT. If N = 1, sample can also be a vector.}
}
\value{
A N x 4 x M array, where N is the sample size, and M is the number of markers.
}
\description{
Obtain allele counts for a sample
}
\details{
get.y(sample)
}
\examples{
get.y(forenseq.snipper.reference$sample)
}
