% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.biallelic.R
\name{is.biallelic}
\alias{is.biallelic}
\title{is.biallelic}
\usage{
is.biallelic(sample)
}
\arguments{
\item{sample}{An N x M matrix with entrie AA,...,TT.}
}
\value{
A binary variable.
}
\description{
A function to check if the data is biallelic
}
\examples{
sample = forenseq.snipper.reference$sample
is.biallelic(sample)
}
