% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.q.R
\name{loglik.q}
\alias{loglik.q}
\title{loglik.q}
\usage{
loglik.q(sample, p, q, sum=TRUE)
}
\arguments{
\item{sample}{A vector of length M giving the sample.}

\item{p}{A K x 4 x M array of allele frequencies.}

\item{q}{A vector of length K of the individual admixture estimates for the sample.}

\item{sum}{A binary variable indicating if the sum over all markers should be reported or all contributions.}
}
\value{
Either a single value (if sum=FALSE) or a vector of length M (if sum=TRUE), giving the contributions to the log-likelihood.
}
\description{
A function to obtain the log-likelihood of the individual admixture of a single sample
}
\examples{
training.sample = forenseq.snipper.reference$sample
training.samLoc = forenseq.snipper.reference$samLoc
test.sample = forenseq.EGDP$sample[1,]
p = get.freqs(training.sample, training.samLoc)
q = get.ia(test.sample, training.sample, training.samLoc)
loglik.q(test.sample, p, q)
}
